Convert from ints and doubles to all floats, the svg native format.
Also create functions for converting model space coordinates to view space

+++ LayoutCAD/ViewModel/BackgroundGridVM.cs
@@ -11,12 +11,12 @@ namespace LayoutCAD.ViewModel
     /// </summary>
     public class BackgroundGridVM
     {
-        public double LineSeparation_mm => 150;
-        public double Width => _viewPort.ViewWidth;
-        public double Height => _viewPort.ViewHeight;
+        public float LineSeparation_mm => 150;
+        public float Width => _viewPort.ViewWidth;
+        public float Height => _viewPort.ViewHeight;
 
         private readonly ViewPort _viewPort;
-        private readonly Func<(double modelCoord, bool isHorizontal), GridLineVM> _lineFactory;
+        private readonly Func<(float modelCoord, bool isHorizontal), GridLineVM> _lineFactory;
 
         public IEnumerable<GridLineVM> GridLines
         {
@@ -24,7 +24,7 @@ namespace LayoutCAD.ViewModel
             {
                 var viewPortModelHeight = _viewPort.ModelSpaceBottom - _viewPort.ModelSpaceTop;
                 var numHorizontalLines = (int)(Math.Ceiling(viewPortModelHeight / LineSeparation_mm));
-                var startPosVertical = Math.Floor(_viewPort.ModelSpaceTop / LineSeparation_mm) * LineSeparation_mm;
+                var startPosVertical = (float)Math.Floor(_viewPort.ModelSpaceTop / LineSeparation_mm) * LineSeparation_mm;
 
                 for (int i = 0; i < numHorizontalLines; i++)
                 {
@@ -33,7 +33,7 @@ namespace LayoutCAD.ViewModel
 
                 var viewPortModelWidth = _viewPort.ModelSpaceRight - _viewPort.ModelSpaceLeft;
                 var numVerticalLines = (int)(Math.Ceiling(viewPortModelWidth / LineSeparation_mm));
-                var startPosHorizontal = Math.Floor(_viewPort.ModelSpaceLeft / LineSeparation_mm) * LineSeparation_mm;
+                var startPosHorizontal = (float)Math.Floor(_viewPort.ModelSpaceLeft / LineSeparation_mm) * LineSeparation_mm;
 
                 for (int i = 0; i < numVerticalLines; i++)
                 {
@@ -44,7 +44,7 @@ namespace LayoutCAD.ViewModel
 
         public BackgroundGridVM(
             ViewPort viewPort,
-            Func<(double modelCoord, bool isHorizontal), GridLineVM> lineFactory)
+            Func<(float modelCoord, bool isHorizontal), GridLineVM> lineFactory)
         {
             _viewPort = viewPort;
             _lineFactory = lineFactory;
